Made styles.js for DrugInEvent
